refactor: update integration tests to use new auth service structure
- Replaced references to the old authService module with the new auth module across LockScreen, LoginScreen, and SignUpScreen tests.
- Updated mock implementations to reflect changes in authentication methods, ensuring tests align with the current API.
- Enhanced test cases for better clarity and accuracy in user authentication scenarios.
